--- a/benchmark/index.js
+++ b/benchmark/index.js
@@ -3,7 +3,7 @@ const bench = require('fastbench')
 const noir = require('pino-noir')(['a.b.c'])
 const fastRedact = require('..')
 
-const censorFn = (v) => v + '.';
+const censorFn = (v) => v + '.'
 
 const redactNoSerialize = fastRedact({ paths: ['a.b.c'], serialize: false })
 const redactWildNoSerialize = fastRedact({ paths: ['a.b.*'], serialize: false })
@@ -28,27 +28,27 @@ const getObj = () => ({
       c: 's'
     }
   }
-});
+})
 
 const max = 500
 
 var run = bench([
   function benchNoirV2 (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       noir.a(obj.a)
     }
     setImmediate(cb)
   },
   function benchFastRedact (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactNoSerialize(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactRestore (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactNoSerialize(obj)
       redactNoSerialize.restore(obj)
@@ -56,21 +56,21 @@ var run = bench([
     setImmediate(cb)
   },
   function benchNoirV2Wild (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       noirWild.a(obj.a)
     }
     setImmediate(cb)
   },
   function benchFastRedactWild (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactWildNoSerialize(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactWildRestore (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactWildNoSerialize(obj)
       redactWildNoSerialize.restore(obj)
@@ -78,14 +78,14 @@ var run = bench([
     setImmediate(cb)
   },
   function benchFastRedactIntermediateWild (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactIntermediateWildNoSerialize(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactIntermediateWildRestore (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactIntermediateWildNoSerialize(obj)
       redactIntermediateWildNoSerialize.restore(obj)
@@ -93,14 +93,14 @@ var run = bench([
     setImmediate(cb)
   },
   function benchJSONStringify (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       JSON.stringify(obj)
     }
     setImmediate(cb)
   },
   function benchNoirV2Serialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       noir.a(obj.a)
       JSON.stringify(obj)
@@ -108,14 +108,14 @@ var run = bench([
     setImmediate(cb)
   },
   function benchFastRedactSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redact(obj)
     }
     setImmediate(cb)
   },
   function benchNoirV2WildSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       noirWild.a(obj.a)
       JSON.stringify(obj)
@@ -123,42 +123,42 @@ var run = bench([
     setImmediate(cb)
   },
   function benchFastRedactWildSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactWild(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactIntermediateWildSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactIntermediateWild(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactIntermediateWildMatchWildOutcomeSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactIntermediateWildMatchWildOutcome(obj)
     }
     setImmediate(cb)
   },
   function benchFastRedactStaticMatchWildOutcomeSerialize (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactStaticMatchWildOutcome(obj)
     }
     setImmediate(cb)
   },
   function benchNoirV2CensorFunction (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       noirCensorFunction.a(obj.a)
     }
     setImmediate(cb)
   },
   function benchFastRedactCensorFunction (cb) {
-    const obj = getObj();
+    const obj = getObj()
     for (var i = 0; i < max; i++) {
       redactCensorFunction(obj)
     }