This creates a new object before each benchmark's loop, which should make the benchmarks more reliable (since hidden class transitions on the object from prior tests could effect the speed of the redaction). This doesn’t go as far as to create a new object right before each redaction call (i.e., in the `for` loop) as that seems like it would add more noise/difficulty interpreting the final results, since a bigger chunk of each benchmark’s work would be this object creation.

This also means that the `serialize: false` tests (and the pino-noir tests, which are effectively serialize false) don’t leave the object in a different state for the subsequent tests, which is required for adding tests for #16. [That's what motivated me to make this change.]

Most of the benchmark times remained similar before and after this change, with some of the fast-redact benchmarks getting 10-30% faster [but idk how much of that is noise]. A few benchmarks seemed to show consistent differences of greater than 30%:

- `benchNoirV2` (the first case): about 50% faster [consistently],  for
reasons I don’t quite understand.

- `benchFastRedactRestore`: about 50% faster

- `benchFastRedactIntermediateWild`: about 50% slower

- `benchNoirV2Wild`,`benchNoirV2CensorFunction`, `benchNoirV2Wild`,
`benchNoirV2CensorFunction`: 2-4x as fast

- `benchFastRedactCensorFunction`: 50-100% faster

Here are the full results:

# Before [ran twice, and the results were pretty similar]

benchNoirV2*500: 84.279ms
benchFastRedact*500: 14.280ms
benchFastRedactRestore*500: 22.773ms
benchNoirV2Wild*500: 105.580ms
benchFastRedactWild*500: 40.533ms
benchFastRedactWildRestore*500: 41.080ms
benchFastRedactIntermediateWild*500: 107.193ms
benchFastRedactIntermediateWildRestore*500: 92.896ms
benchJSONStringify*500: 323.536ms
benchNoirV2Serialize*500: 407.667ms
benchFastRedactSerialize*500: 337.723ms
benchNoirV2WildSerialize*500: 380.418ms
benchFastRedactWildSerialize*500: 372.057ms
benchFastRedactIntermediateWildSerialize*500: 417.458ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 572.464ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 348.632ms
benchNoirV2CensorFunction*500: 70.804ms
benchFastRedactCensorFunction*500: 59.476ms
benchNoirV2*500: 48.808ms
benchFastRedact*500: 11.550ms
benchFastRedactRestore*500: 13.436ms
benchNoirV2Wild*500: 61.383ms
benchFastRedactWild*500: 31.472ms
benchFastRedactWildRestore*500: 34.325ms
benchFastRedactIntermediateWild*500: 128.071ms
benchFastRedactIntermediateWildRestore*500: 128.602ms
benchJSONStringify*500: 317.474ms
benchNoirV2Serialize*500: 395.179ms
benchFastRedactSerialize*500: 323.449ms
benchNoirV2WildSerialize*500: 369.051ms
benchFastRedactWildSerialize*500: 356.204ms
benchFastRedactIntermediateWildSerialize*500: 426.994ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 531.789ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 340.540ms
benchNoirV2CensorFunction*500: 38.953ms
benchFastRedactCensorFunction*500: 53.361ms

# After [ran twice, and the results were pretty similar]

benchNoirV2*500: 54.057ms
benchFastRedact*500: 15.949ms
benchFastRedactRestore*500: 13.749ms
benchNoirV2Wild*500: 34.787ms
benchFastRedactWild*500: 44.642ms
benchFastRedactWildRestore*500: 42.949ms
benchFastRedactIntermediateWild*500: 162.311ms
benchFastRedactIntermediateWildRestore*500: 106.921ms
benchJSONStringify*500: 306.766ms
benchNoirV2Serialize*500: 403.440ms
benchFastRedactSerialize*500: 321.294ms
benchNoirV2WildSerialize*500: 366.650ms
benchFastRedactWildSerialize*500: 342.946ms
benchFastRedactIntermediateWildSerialize*500: 437.797ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 573.556ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 336.521ms
benchNoirV2CensorFunction*500: 27.110ms
benchFastRedactCensorFunction*500: 47.454ms
benchNoirV2*500: 38.437ms
benchFastRedact*500: 10.272ms
benchFastRedactRestore*500: 9.693ms
benchNoirV2Wild*500: 18.504ms
benchFastRedactWild*500: 30.266ms
benchFastRedactWildRestore*500: 35.108ms
benchFastRedactIntermediateWild*500: 131.794ms
benchFastRedactIntermediateWildRestore*500: 110.691ms
benchJSONStringify*500: 299.861ms
benchNoirV2Serialize*500: 384.236ms
benchFastRedactSerialize*500: 314.049ms
benchNoirV2WildSerialize*500: 365.485ms
benchFastRedactWildSerialize*500: 344.158ms
benchFastRedactIntermediateWildSerialize*500: 426.421ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 537.079ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 340.104ms
benchNoirV2CensorFunction*500: 16.021ms
benchFastRedactCensorFunction*500: 31.100ms