@davidmarkclements 2019-10-09
have you verified that V8 doesn't optimize for the `getObj` shape (which would essentially lead to the same situation)? If we could randomly generate slightly different shapes (but with the necessary paths) this would be even more realistic.