@ethanresnick 2019-10-09
> have you verified that V8 doesn't optimize for the getObj shape (which would essentially lead to the same situation)?

I haven't verified anything in v8...but I'm not sure I understand the point you're raising. What I'm trying to control with this PR is the object's shape was previously changing between benchmark cases: it starts with the keys as strings, then pino-noir sets some of the keys to [`Redacted`](https://github.com/pinojs/pino-noir/blob/master/index.js#L169) objects, and then default `censor` value for `fast-redact` sets them back to strings, and so on.