@ethanresnick 2019-10-09
In the meantime, I've fixed up the style