@davidmarkclements 2019-10-09
Yeah that’s true this does remove the confounding factor of the shared obj 

V8 also uses hidden classes (aka shapes, aka “maps”) to optimise multiple objects created with the same shape. So that means the function returning the same shape can also suffer from a related issue where the later tests will be faster because of behind the scenes optimisation. To avoid this varying the shape between scenarios  should  be even more realistic