@ethanresnick 2019-10-10
> V8 also uses hidden classes (aka shapes, aka “maps”) to optimise multiple objects created with the same shape. So that means the function returning the same shape can also suffer from a related issue where the later tests will be faster because of behind the scenes optimisation. To avoid this varying the shape between scenarios should be even more realistic

Ahh, gotcha. And I’m not opposed to randomly tweaking the shape per test. 

Two ideas, though, which might be simpler:

1. Maybe we could just move the `getObj()` calls outside the benchmark case (i.e., make each case’s object before we start the clock)? I guess we’d store those objects in an array, and each benchmark case would start with `const obj = objects[xxxx]`, since I think(?) it’s safe to assume that lookup will always be the same speed?

2. If the above assumption isn’t safe, maybe we can just call `getObj` a bunch of times in a loop [where the result’s saved/used somewhere so the whole loop isn’t compiled away], to get V8 into kicking in whatever optimizations it’s gonna make for `getObj` before any test start.

That said, the object creation only happens once per case, vs the redaction happening 500 times, so I have to imagine that that creation isn’t effecting the results much