@davidmarkclements 2019-10-10
yeah creation isn't an issue - but rather it's possible creating the same shape a few times can trigger an optimization of the shape.

Since theres only a few cases you could just create a different object (but equal in size) for each benchmark - so for instance, create different single little properties in each case