@ethanresnick 2019-10-11
Not really, or at least it all looks like noise to me:

```
benchNoirV2*500: 56.187ms
benchFastRedact*500: 15.636ms
benchFastRedactRestore*500: 18.708ms
benchNoirV2Wild*500: 37.492ms
benchFastRedactWild*500: 41.489ms
benchFastRedactWildRestore*500: 41.012ms
benchFastRedactIntermediateWild*500: 177.360ms
benchFastRedactIntermediateWildRestore*500: 99.931ms
benchJSONStringify*500: 321.774ms
benchNoirV2Serialize*500: 406.218ms
benchFastRedactSerialize*500: 324.288ms
benchNoirV2WildSerialize*500: 419.127ms
benchFastRedactWildSerialize*500: 369.263ms
benchFastRedactIntermediateWildSerialize*500: 450.892ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 661.560ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 375.092ms
benchNoirV2CensorFunction*500: 26.578ms
benchFastRedactCensorFunction*500: 41.205ms
benchNoirV2*500: 39.721ms
benchFastRedact*500: 8.976ms
benchFastRedactRestore*500: 10.869ms
benchNoirV2Wild*500: 16.557ms
benchFastRedactWild*500: 27.925ms
benchFastRedactWildRestore*500: 32.596ms
benchFastRedactIntermediateWild*500: 172.279ms
benchFastRedactIntermediateWildRestore*500: 125.881ms
benchJSONStringify*500: 312.815ms
benchNoirV2Serialize*500: 406.060ms
benchFastRedactSerialize*500: 333.660ms
benchNoirV2WildSerialize*500: 419.161ms
benchFastRedactWildSerialize*500: 365.389ms
benchFastRedactIntermediateWildSerialize*500: 452.600ms
benchFastRedactIntermediateWildMatchWildOutcomeSerialize*500: 630.983ms
benchFastRedactStaticMatchWildOutcomeSerialize*500: 370.881ms
benchNoirV2CensorFunction*500: 16.885ms
benchFastRedactCensorFunction*500: 29.958ms
```